pyproject.toml
[tool.poetry]
name = "test-parser-bse-gui-poetry"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyserial = "^3.5"
pyqt6 = "^6.7.0"
pyqtgraph = "^0.13.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
